Add api for subscription pre-condition checking
The purpose of this API is to sure that the system is ready to consume
events. For example if the events consumed would end up in solr indexes
then Solr being up and running should be a pre-condition to check before
consuming the events. If pre-condition is not met then subscription will
be postponed for next trigger.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>