Store the URI error occurred during consumption
This fix ensures that in case an error occurs the event has to be deleted
manually. The reason is, "consumer" should never throw any exception that
it can not handle at a later point of time without manual intervention. If the
event message is in incorrect format it should simply ignore it and move
ahead and not throw exception. A usecase of "consumer" throwing an
exception would be if Apache Solr is not available.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>